+++ lib/src/core/application/services/init.dart
@@ -42,7 +42,6 @@ class Init {
     await dir.create(recursive: true);
     final dbPath = join(dir.path, 'lab_ferreira_sembast.db');
     final isar = await Isar.open(
-        directory: dir.path,
         [ContatoStorageModelSchema, AccountStorageModelSchema],
         inspector: true);
     return await databaseFactoryIo.openDatabase(dbPath);
