+++ pubspec.yaml
@@ -8,15 +8,15 @@ environment:
   sdk: '>=2.18.4 <3.0.0'
 
 dependencies:
-  collection: ^1.16.0
+  collection: ^1.17.0
   connectivity_plus: ^3.0.2
   cupertino_icons: ^1.0.5
   dio: ^4.0.6
   email_validator: ^2.1.17
-  firebase_core: ^2.1.1
-  firebase_crashlytics: ^3.0.4
-  firebase_performance: ^0.9.0+4
-  firebase_storage: ^11.0.4
+  firebase_core: ^2.2.0
+  firebase_crashlytics: ^3.0.5
+  firebase_performance: ^0.9.0+5
+  firebase_storage: ^11.0.5
   flex_color_scheme: ^6.0.1
   flex_seed_scheme: ^1.0.1
   flutter:
@@ -26,7 +26,7 @@ dependencies:
   flutter_riverpod: ^2.1.1
   formz: ^0.4.1
   freezed_annotation: ^2.2.0
-  go_router: ^5.1.5
+  go_router: ^5.1.6
   isar: ^3.0.4
   isar_flutter_libs: ^3.0.4
   json_annotation: ^4.7.0
@@ -37,18 +37,22 @@ dependencies:
   path_provider: ^2.0.11
   sembast: ^3.3.1
 
-dev_dependencies:
+dependency_overrides:
   build_runner: ^2.3.0
+  freezed: ^2.2.0
+  collection: ^1.16.0
+
+dev_dependencies:
+  build_runner: ^2.3.2
   faker: ^2.0.0
   flutter_lints: ^2.0.1
   flutter_native_splash: ^2.2.14
   flutter_test:
     sdk: flutter
-  freezed: ^2.2.0
+  freezed: ^2.2.1
   isar_generator: ^3.0.4
   json_serializable: ^6.5.4
   mocktail: ^0.3.0
-  source_gen: ^1.2.6
 
 flutter_native_splash:
   image: assets/laboratorioferreira_logo_sem_fundo.png
@@ -69,7 +73,8 @@ scripts:
     - flutter clean
     - flutter pub get
     - flutter pub upgrade
-  test: flutter test --verbose
+    - derry gen
+  test: flutter test --coverage
   update:
     - flutter pub get
     - flutter pub upgrade
