+++ test/src/features/auth/data/repositories/auth_repository_test.dart
@@ -1,16 +1,16 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:laboratorio_ferreira_mobile/src/core/core.dart';
-import 'package:laboratorio_ferreira_mobile/src/features/auth/auth.dart';
+import 'package:laboratorio_ferreira_mobile/src/core/application/application.dart';
+import 'package:laboratorio_ferreira_mobile/src/features/auth/data/repositories/auth_repository.dart';
 import 'package:mocktail/mocktail.dart';
 
-class DioServiceMock extends Mock implements DioService {}
+class HttpServiceMock extends Mock implements IHttpService {}
 
 void main() {
-  late final DioServiceMock dioServiceMock;
+  late final HttpServiceMock httpServiceMock;
   late final AuthRepository authRepository;
 
   setUpAll(() {
-    dioServiceMock = DioServiceMock();
-    authRepository = AuthRepository(httpService: dioServiceMock);
+    httpServiceMock = HttpServiceMock();
+    // authRepository = AuthRepository(httpService: httpServiceMock);
   });
 }
